requirements.txt
pip
google-cloud-datastore
google-cloud-tasks
pytest